refactor: split callback_handlers.py (2216 lines) into 7 domain modules
